backend/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
fastapi>=0.100.0
httpx>=0.23.0
langchain-core>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
motor>=3.3.0
passlib[bcrypt]>=1.7.4
Pillow>=9.0.0
protobuf>=4.21.0
pydantic[email]>=2.0.0
pymongo>=4.5.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
redis>=5.0.0
requests>=2.31.0
selenium>=4.10.0
uvicorn>=0.22.0
webdriver-manager>=4.0.0
google-genai>=0.3.0
langgraph-checkpoint-mongodb>=0.1.0
python-multipart>=0.0.6
bcrypt==3.1.7
email-validator>=2.0.0
